Row bind list of data frames wrapper to one-function do.call rbind over a lapply list

Row bind list of data frames wrapper to one-function do.call rbind over a lapply list

rbind_lapply(X, FUN)

Arguments

X
a list. See details lapply.

FUN
a function to subset data frames (or data tables). See details lapply.

Value

a row bound data frame

Examples

set.seed(42)

X<- lapply(1:3,function(i) list(int = 42,df = data.frame(x = runif(10,0,1),y = runif(10,0,1))))
rbind_lapply(X,function(x) x$df)

#>             x           y
#> 1  0.91480604 0.457741776
#> 2  0.93707541 0.719112252
#> 3  0.28613953 0.934672247
#> 4  0.83044763 0.255428824
#> 5  0.64174552 0.462292823
#> 6  0.51909595 0.940014523
#> 7  0.73658831 0.978226428
#> 8  0.13466660 0.117487362
#> 9  0.65699229 0.474997082
#> 10 0.70506478 0.560332746
#> 11 0.90403139 0.737595618
#> 12 0.13871017 0.811055141
#> 13 0.98889173 0.388108283
#> 14 0.94666823 0.685169729
#> 15 0.08243756 0.003948339
#> 16 0.51421178 0.832916080
#> 17 0.39020347 0.007334147
#> 18 0.90573813 0.207658973
#> 19 0.44696963 0.906601408
#> 20 0.83600426 0.611778643
#> 21 0.37955924 0.333427211
#> 22 0.43577158 0.346748248
#> 23 0.03743103 0.398485411
#> 24 0.97353991 0.784692776
#> 25 0.43175125 0.038936491
#> 26 0.95757660 0.748795386
#> 27 0.88775491 0.677276830
#> 28 0.63997877 0.171264330
#> 29 0.97096661 0.261087964
#> 30 0.61883821 0.514412935